Add Implicit Casting

+++ tests/alltypesDeclarations_converted.rs
@@ -1,5 +1,5 @@
 #![allow(warnings, unused)]
 use std::*;
 fn main() {
-    const hello: i32 = 5;
+    const hello: i32 = 5 as i32;
 }
