+++ tests/casting_converted.rs
@@ -1,9 +1,9 @@
 #![allow(warnings, unused)]
 use std::*;
 fn main() {
-    let mut x: i32 = 5;
-    let mut t: i32 = x as i32;
-    let mut a: i64 = x as i64;
-    let mut l: u64 = x as u64;
-    let mut l: i32 = (x as i32) as i32;
+    let mut x: i32 = 5 as i32;
+    let mut t: i32 = x as i32 as i32;
+    let mut a: i64 = x as i64 as i64;
+    let mut l: u64 = x as u64 as u64;
+    let mut l: i32 = (x as i32) as i32 as i32;
 }
